// WADNR.Web/src/app/shared/generated/model/fund-source-allocation-program-index-project-code-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FundSourceAllocationProgramIndexProjectCodePair } from './fund-source-allocation-program-index-project-code-pair';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FundSourceAllocationProgramIndexProjectCodeSaveRequest { 
    Pairs?: Array<FundSourceAllocationProgramIndexProjectCodePair> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FundSourceAllocationProgramIndexProjectCodeSaveRequestForm { 
    Pairs?: FormControl<Array<FundSourceAllocationProgramIndexProjectCodePair>>;
}

export class FundSourceAllocationProgramIndexProjectCodeSaveRequestFormControls { 
    public static Pairs = (value: FormControlState<Array<FundSourceAllocationProgramIndexProjectCodePair>> | Array<FundSourceAllocationProgramIndexProjectCodePair> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<FundSourceAllocationProgramIndexProjectCodePair>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}